fix: issue where tailwind style not applied directly
